Only use own translator visitor
