+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+1.0.2
+=====
+
+*   (bug) Only use own translator visitor to reduce number of false-positives.
+
+
 1.0.1
 =====
